+++ checkpoint/orbax/checkpoint/experimental/v1/_src/layout/numpy_layout_test.py
@@ -93,9 +93,8 @@ class NumpyLayoutTest(unittest.IsolatedAsyncioTestCase, parameterized.TestCase):
 
     # Load the checkpoint
     layout = NumpyLayout(test_path)
-    restore_fn = await layout.load()
-    restored_checkpointables = await restore_fn
-    pytree = restored_checkpointables['pytree']
+    restore_fn = await layout.load_pytree()
+    pytree = await restore_fn
 
     # Verify restored data
     if np.issubdtype(dtype, np.floating):