+++ checkpoint/orbax/checkpoint/experimental/v1/_src/layout/pytorch_layout.py
@@ -308,7 +308,7 @@ def _unpickle_structure(
 async def _load_pytorch(
     path: Path,
     abstract_pytree: tree_types.PyTreeOf[jax.ShapeDtypeStruct] | None = None,
-) -> dict[str, Any]:
+) -> Any:
   """Loads pytorch checkpoint as numpy arrays or sharded jax arrays."""
   pickle_bytes, storage_data = await _read_zip_contents(path)
 
@@ -321,7 +321,7 @@ async def _load_pytorch(
     # Return on-device JAX arrays.
     restored_pytree = _load_pytorch_on_device(pytorch_data, abstract_pytree)
 
-  return {checkpoint_layout.PYTREE_CHECKPOINTABLE_KEY: restored_pytree}
+  return restored_pytree
 
 
 class PyTorchLayout(CheckpointLayout):
@@ -389,7 +389,16 @@ class PyTorchLayout(CheckpointLayout):
         commit_timestamp_nsecs=commit_timestamp_nsecs,
     )
 
-  async def load(
+  async def load_pytree(
+      self,
+      checkpointable_name: str | None = None,
+      abstract_pytree: Any | None = None,
+  ) -> Awaitable[Any]:
+    del checkpointable_name
+    load_awaitable = _load_pytorch(self._path, abstract_pytree)
+    return load_awaitable
+
+  async def load_checkpointables(
       self,
       abstract_checkpointables: (
           dict[str, tree_types.PyTreeOf[jax.ShapeDtypeStruct]] | None
@@ -413,4 +422,10 @@ class PyTorchLayout(CheckpointLayout):
       abstract_pytree = abstract_checkpointables.get(
           checkpoint_layout.PYTREE_CHECKPOINTABLE_KEY
       )
-    return _load_pytorch(self._path, abstract_pytree)
+
+    async def _loader():
+      restored_pytree = await _load_pytorch(self._path, abstract_pytree)
+      return {checkpoint_layout.PYTREE_CHECKPOINTABLE_KEY: restored_pytree}
+
+    load_awaitable = _loader()
+    return load_awaitable