+++ checkpoint/orbax/checkpoint/experimental/v1/_src/metadata/loading.py
@@ -89,17 +89,23 @@ def pytree_metadata(
   asyncio_utils.maybe_apply_nest_asyncio()
   ctx = context_lib.get_context()
   path = ctx.file_options.path_class(path)
+
   layout, checkpointable_name = asyncio.run(
       layout_registry.get_checkpoint_layout_pytree(
           path, ctx.checkpoint_layout, checkpointable_name
       )
   )
-  metadata = _checkpointables_metadata_impl(layout)
+
+  step_metadata = _checkpointables_metadata_impl(layout)
+  if checkpointable_name is None:
+    metadata = step_metadata.metadata
+  else:
+    metadata = step_metadata.metadata[checkpointable_name]
   return CheckpointMetadata[PyTreeMetadata](
-      metadata=metadata.metadata[checkpointable_name],
-      init_timestamp_nsecs=metadata.init_timestamp_nsecs,
-      commit_timestamp_nsecs=metadata.commit_timestamp_nsecs,
-      custom_metadata=metadata.custom_metadata,
+      metadata=metadata,
+      init_timestamp_nsecs=step_metadata.init_timestamp_nsecs,
+      commit_timestamp_nsecs=step_metadata.commit_timestamp_nsecs,
+      custom_metadata=step_metadata.custom_metadata,
   )