+++ checkpoint/CHANGELOG.md
@@ -18,6 +18,13 @@ creation scheduling, which may slow down background saving
 (but not the blocking time). The change also improves the
 `PathAwaitingCreation` interface so that any call to `await_creation` will
 trigger creation if it has not already been started.
+- Add new `OrbaxV0Layout` that will handle specific v0 checkpoint format logic.
+- Refactor `CheckpointLayout` splitting `load()` into `load_pytree()` and
+`load_checkpointables()` each with their own dedicated loading logic
+- Refactor v0 Pytree validation and metadata resolution and add `OrbaxV0Layout`
+tests
+- Fix composite handler's extraction logic for handler type strings when
+missing `_CHECKPOINT_METADATA`
 
 ## [0.11.31] - 2025-12-11